requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
plotly>=5.20.0
pandas>=2.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
